+++ R/assign_summary_digits.R
@@ -114,9 +114,9 @@ assign_summary_digits <- function(data, statistic, type, digits = NULL) {
       if (!is_integerish(value)) {
         return(value)
       }
-      # if an integer is passed for a percentage, process stat with style_percent()
+      # if an integer is passed for a percentage, process stat with label_style_number()
       if (stat_name %in% c("p", "p_miss", "p_nonmiss", "p_unweighted")) {
-        return(label_style_percent(digits = value))
+        return(label_style_number(digits = value, scale = 100))
       }
       # otherwise, use style_number() to style number
       return(label_style_number(digits = value))